+++ magick/zoom.c
@@ -962,7 +962,8 @@ static unsigned int VerticalFilter(const Image *source,Image *destination,
       opacity=0.0;
       for (i=0; i < n; i++)
       {
-        j=(contribution[i].pixel-contribution[0].pixel)*source->columns+x;
+        j=(long) (contribution[i].pixel-contribution[0].pixel)*
+          source->columns+x;
         red+=contribution[i].weight*(p+j)->red;
         green+=contribution[i].weight*(p+j)->green;
         blue+=contribution[i].weight*(p+j)->blue;
@@ -970,7 +971,8 @@ static unsigned int VerticalFilter(const Image *source,Image *destination,
       }
       if (indexes != (IndexPacket *) NULL)
         {
-          j=(contribution[n/2].pixel-contribution[0].pixel)*source->columns+x;
+          j=(long) (contribution[n/2].pixel-contribution[0].pixel)*
+            source->columns+x;
           indexes[x]=(GetIndexes(source))[j];
         }
       q->red=(Quantum) ((red < 0) ? 0 : (red > MaxRGB) ? MaxRGB : red+0.5);
